[system] sleep: splits system_sleep.cpp/h into several files to more easily manage
