+++ system/src/system_sleep.cpp
@@ -19,217 +19,20 @@
 
 
 #include "system_sleep.h"
+#include "system_sleep_configuration.h"
 #include "system_network.h"
-#include "system_task.h"
+#include "system_power.h"
+#include "system_threading.h"
 #include "system_cloud.h"
 #include "system_cloud_internal.h"
-#include "system_network_internal.h"
-#include "system_threading.h"
-#include "rtc_hal.h"
-#include "core_hal.h"
-#include "led_service.h"
-#include <stddef.h>
-#include "spark_wiring_fuel.h"
+#include "system_mode.h"
 #include "spark_wiring_system.h"
-#include "spark_wiring_platform.h"
-#include "system_power.h"
+#include "led_service.h"
+#if HAL_PLATFORM_CELLULAR
+#include "cellular_hal.h"
+#endif // HAL_PLATFORM_CELLULAR
 #include "check.h"
 
-#if PLATFORM_ID==PLATFORM_ELECTRON_PRODUCTION
-# include "parser.h"
-#endif
-
-
-struct WakeupState
-{
-    bool wifi;
-    bool wifiConnected;
-    bool cloud;
-};
-
-WakeupState wakeupState;
-
-static void network_suspend() {
-    // save the current state so it can be restored on wakeup
-
-    wakeupState.cloud = spark_cloud_flag_auto_connect();
-    wakeupState.wifi = !SPARK_WLAN_SLEEP;
-    wakeupState.wifiConnected = wakeupState.cloud || network_ready(0, 0, NULL) || network_connecting(0, 0, NULL);
-    // Disconnect the cloud and the network
-    network_disconnect(0, NETWORK_DISCONNECT_REASON_SLEEP, NULL);
-    // Clear the auto connect status
-    spark_cloud_flag_disconnect();
-    network_off(0, 0, 0, NULL);
-}
-
-static void network_resume() {
-    // Set the system flags that triggers the wifi/cloud reconnection in the background loop
-    if (wakeupState.wifiConnected || wakeupState.wifi)  // at present, no way to get the background loop to only turn on wifi.
-        SPARK_WLAN_SLEEP = 0;
-    if (wakeupState.cloud)
-        spark_cloud_flag_connect();
-}
-
-/*******************************************************************************
- * Function Name  : HAL_RTCAlarm_Handler
- * Description    : This function handles additional application requirements.
- * Input          : None.
- * Output         : None.
- * Return         : None.
- *******************************************************************************/
-extern "C" void HAL_RTCAlarm_Handler(void)
-{
-    /* Wake up from System.sleep mode(SLEEP_MODE_WLAN) */
-    network_resume();
-}
-
-void sleep_fuel_gauge()
-{
-    FuelGauge gauge;
-
-    gauge.sleep();
-}
-
-bool network_sleep_flag(uint32_t flags)
-{
-    return (flags & SYSTEM_SLEEP_FLAG_NETWORK_STANDBY) == 0;
-}
-
-int system_sleep_impl(Spark_Sleep_TypeDef sleepMode, long seconds, uint32_t param, void* reserved)
-{
-    SYSTEM_THREAD_CONTEXT_SYNC(system_sleep_impl(sleepMode, seconds, param, reserved));
-    // TODO - determine if these are valuable:
-    // - Currently publishes will get through with or without #1.
-    // - More data is consumed with #1.
-    // - Session is not resuming after waking from DEEP sleep,
-    //   so a full handshake currently outweighs leaving the
-    //   modem on for #2.
-    //
-    //---- #1
-    // If we're connected to the cloud, make sure all
-    // confirmable UDP messages are sent before sleeping
-    // if (spark_cloud_flag_connected()) {
-    //     Spark_Sleep();
-    // }
-    //---- #2
-    // SLEEP_NETWORK_STANDBY can keep the modem on during DEEP sleep
-    // System.sleep(10) always powers down the network, even if SLEEP_NETWORK_STANDBY flag is used.
-
-    // Make sure all confirmable UDP messages are sent and acknowledged before sleeping
-    if (spark_cloud_flag_connected() && !(param & SYSTEM_SLEEP_FLAG_NO_WAIT)) {
-        Spark_Sleep();
-    }
-
-    if (network_sleep_flag(param) || SLEEP_MODE_WLAN == sleepMode) {
-        network_suspend();
-    }
-
-    switch (sleepMode)
-    {
-        case SLEEP_MODE_WLAN:
-            if (seconds)
-            {
-                HAL_RTC_Set_UnixAlarm((time_t) seconds);
-            }
-            break;
-
-        case SLEEP_MODE_DEEP:
-            if (network_sleep_flag(param))
-            {
-                network_disconnect(0, NETWORK_DISCONNECT_REASON_SLEEP, NULL);
-                network_off(0, 0, 0, NULL);
-            }
-
-            system_power_management_sleep();
-            return HAL_Core_Enter_Standby_Mode(seconds,
-                    (param & SLEEP_DISABLE_WKP_PIN.value()) ? HAL_STANDBY_MODE_FLAG_DISABLE_WKP_PIN : 0);
-            break;
-
-#if HAL_PLATFORM_SETUP_BUTTON_UX
-        case SLEEP_MODE_SOFTPOWEROFF:
-            network_disconnect(0, NETWORK_DISCONNECT_REASON_SLEEP, NULL);
-            network_off(0, 0, 0, NULL);
-            sleep_fuel_gauge();
-            system_power_management_sleep();
-            return HAL_Core_Enter_Standby_Mode(seconds,
-                    (param & SLEEP_DISABLE_WKP_PIN.value()) ? HAL_STANDBY_MODE_FLAG_DISABLE_WKP_PIN : 0);
-            break;
-#endif
-    }
-    return 0;
-}
-
-int system_sleep_pin_impl(const uint16_t* pins, size_t pins_count, const InterruptMode* modes, size_t modes_count, long seconds, uint32_t param, void* reserved)
-{
-    SYSTEM_THREAD_CONTEXT_SYNC(system_sleep_pin_impl(pins, pins_count, modes, modes_count, seconds, param, reserved));
-
-    // Make sure all confirmable UDP messages are sent and acknowledged before sleeping
-    if (spark_cloud_flag_connected() && !(param & SYSTEM_SLEEP_FLAG_NO_WAIT)) {
-        Spark_Sleep();
-    }
-
-    bool network_sleep = network_sleep_flag(param);
-    if (network_sleep)
-    {
-        network_suspend();
-    }
-
-#if HAL_PLATFORM_CELLULAR
-    if (!network_sleep_flag(param)) {
-        // Pause the modem Serial
-        cellular_pause(nullptr);
-    }
-#endif // HAL_PLATFORM_CELLULAR
-
-#if HAL_PLATFORM_MESH
-    // FIXME: We are still going to turn off OpenThread with SLEEP_NETWORK_STANDBY, otherwise
-    // there are various issues with sleep
-    if (!network_sleep_flag(param)) {
-        network_off(NETWORK_INTERFACE_MESH, 0, 0, nullptr);
-    }
-#endif // HAL_PLATFORM_MESH
-
-    led_set_update_enabled(0, nullptr); // Disable background LED updates
-    LED_Off(LED_RGB);
-	system_power_management_sleep();
-    int ret = HAL_Core_Enter_Stop_Mode_Ext(pins, pins_count, modes, modes_count, seconds, nullptr);
-    led_set_update_enabled(1, nullptr); // Enable background LED updates
-
-#if HAL_PLATFORM_CELLULAR
-    if (!network_sleep_flag(param)) {
-        // Pause the modem Serial
-        cellular_resume(nullptr);
-    }
-#endif // HAL_PLATFORM_CELLULAR
-
-#if HAL_PLATFORM_MESH
-    // FIXME: we need to bring Mesh interface back up because we've turned it off
-    // despite SLEEP_NETWORK_STANDBY
-    if (!network_sleep_flag(param)) {
-        network_on(NETWORK_INTERFACE_MESH, 0, 0, nullptr);
-    }
-#endif // HAL_PLATFORM_MESH
-
-    if (network_sleep)
-    {
-        network_resume();   // asynchronously bring up the network/cloud
-    }
-
-    // if single-threaded, managed mode then reconnect to the cloud (for up to 60 seconds)
-    auto mode = system_mode();
-    if (system_thread_get_state(nullptr)==spark::feature::DISABLED && (mode==AUTOMATIC || mode==SEMI_AUTOMATIC) && spark_cloud_flag_auto_connect()) {
-        waitFor(spark_cloud_flag_connected, 60000);
-    }
-
-    if (spark_cloud_flag_connected()) {
-        Spark_Wake();
-    }
-    return ret;
-}
-
-#if HAL_PLATFORM_SLEEP20
-
-#if HAL_PLATFORM_CELLULAR || HAL_PLATFORM_WIFI || HAL_PLATFORM_MESH || HAL_PLATFORM_ETHERNET
 static bool system_sleep_network_suspend(network_interface_index index) {
     bool resume = false;
     // Disconnect from network
@@ -250,7 +53,6 @@ static int system_sleep_network_resume(network_interface_index index) {
     network_connect(index, 0, 0, nullptr);
     return SYSTEM_ERROR_NONE;
 }
-#endif // HAL_PLATFORM_CELLULAR || HAL_PLATFORM_WIFI || HAL_PLATFORM_MESH || HAL_PLATFORM_ETHERNET
 
 int system_sleep_ext(const hal_sleep_config_t* config, hal_wakeup_source_base_t** reason, void* reserved) {
     SYSTEM_THREAD_CONTEXT_SYNC(system_sleep_ext(config, reason, reserved));
@@ -260,21 +62,18 @@ int system_sleep_ext(const hal_sleep_config_t* config, hal_wakeup_source_base_t*
     CHECK(hal_sleep_validate_config(config, nullptr));
 
     SystemSleepConfigurationHelper configHelper(config);
-    int ret;
 
-#ifndef SPARK_NO_CLOUD
     bool cloudResume = false;
     // Disconnect from cloud is necessary.
     // Make sure all confirmable UDP messages are sent and acknowledged before sleeping
     if (configHelper.cloudDisconnectRequested() && spark_cloud_flag_connected()) {
-        if (configHelper.sleepWait() == SystemSleepWait::CLOUD) {
+        if (configHelper.sleepFlags().isSet(SystemSleepFlag::WAIT_CLOUD)) {
             Spark_Sleep();
         }
         cloudResume = spark_cloud_flag_auto_connect();
         // Clear the auto connect status
         spark_cloud_flag_disconnect();
     }
-#endif // SPARK_NO_CLOUD
 
     // Network disconnect.
     // FIXME: if_get_list() can be potentially used, instead of using pre-processor.
@@ -326,7 +125,7 @@ int system_sleep_ext(const hal_sleep_config_t* config, hal_wakeup_source_base_t*
     system_power_management_sleep();
 
     // Now enter sleep mode
-    ret = hal_sleep_enter(config, reason, nullptr);
+    int ret = hal_sleep_enter(config, reason, nullptr);
 
     system_power_management_sleep(false);
 
@@ -364,7 +163,6 @@ int system_sleep_ext(const hal_sleep_config_t* config, hal_wakeup_source_base_t*
     }
 #endif // HAL_PLATFORM_ETHERNET
 
-#ifndef SPARK_NO_CLOUD
     if (cloudResume) {
         // Resume cloud connection.
         spark_cloud_flag_connect();
@@ -378,135 +176,6 @@ int system_sleep_ext(const hal_sleep_config_t* config, hal_wakeup_source_base_t*
             Spark_Wake();
         }
     }
-#endif // SPARK_NO_CLOUD
 
     return ret;
-}
-
-#endif // HAL_PLATFORM_SLEEP20
-
-/**
- * Wraps the actual implementation, which has to return a value as part of the threaded implementation.
- */
-int system_sleep_pin(uint16_t wakeUpPin, uint16_t edgeTriggerMode, long seconds, uint32_t param, void* reserved)
-{
-    // Cancel current connection attempt to unblock the system thread
-    network_connect_cancel(0, 1, 0, 0);
-    InterruptMode m = (InterruptMode)edgeTriggerMode;
-    return system_sleep_pins(&wakeUpPin, 1, &m, 1, seconds, param, reserved);
-}
-
-int system_sleep(Spark_Sleep_TypeDef sleepMode, long seconds, uint32_t param, void* reserved)
-{
-#if HAL_PLATFORM_SLEEP20
-    if (sleepMode == SLEEP_MODE_WLAN) {
-        network_connect_cancel(0, 1, 0, 0);
-        return system_sleep_impl(sleepMode, seconds, param, reserved);
-    } else {
-        SystemSleepConfiguration config;
-        // For backward compatibility. This API will make device enter HIBERNATE mode.
-        config.mode(SystemSleepMode::HIBERNATE);
-
-        if (seconds > 0) {
-            config.duration(seconds * 1000);
-        }
-
-        if (sleepMode == SLEEP_MODE_DEEP && (param & SYSTEM_SLEEP_FLAG_NETWORK_STANDBY)) {
-            // FIXME: if_get_list() can be potentially used, instead of using pre-processor.
-#if HAL_PLATFORM_CELLULAR
-            config.network(NETWORK_INTERFACE_CELLULAR);
-#endif // HAL_PLATFORM_CELLULAR
-
-#if HAL_PLATFORM_WIFI
-            config.network(NETWORK_INTERFACE_WIFI_STA);
-#endif // HAL_PLATFORM_WIFI
-
-#if HAL_PLATFORM_MESH
-            config.network(NETWORK_INTERFACE_MESH);
-#endif // HAL_PLATFORM_MESH
-
-#if HAL_PLATFORM_ETHERNET
-            config.network(NETWORK_INTERFACE_ETHERNET);
-#endif // HAL_PLATFORM_ETHERNET
-        }
-
-        if (param & SYSTEM_SLEEP_FLAG_NO_WAIT) {
-            config.wait(SystemSleepWait::NO_WAIT);
-        }
-
-        if (!(param & SYSTEM_SLEEP_FLAG_DISABLE_WKP_PIN)) {
-            config.gpio(WKP, RISING);
-        }
-
-        return system_sleep_ext(config.halConfig(), nullptr, nullptr);
-    }
-#else
-    network_connect_cancel(0, 1, 0, 0);
-    return system_sleep_impl(sleepMode, seconds, param, reserved);
-#endif // HAL_PLATFORM_SLEEP20
-}
-
-int system_sleep_pins(const uint16_t* pins, size_t pins_count, const InterruptMode* modes, size_t modes_count, long seconds, uint32_t param, void* reserved)
-{
-#if HAL_PLATFORM_SLEEP20
-    SystemSleepConfiguration config;
-    // For backward compatibility. This API will make device enter STOP mode.
-    config.mode(SystemSleepMode::STOP);
-
-    if (seconds > 0) {
-        config.duration(seconds * 1000);
-    }
-
-    for (size_t i = 0; i < pins_count; i++) {
-        config.gpio(pins[i], ((i < modes_count) ? modes[i] : modes[modes_count - 1]));
-    }
-
-    if (param & SYSTEM_SLEEP_FLAG_NETWORK_STANDBY) {
-        // FIXME: if_get_list() can be potentially used, instead of using pre-processor.
-#if HAL_PLATFORM_CELLULAR
-        config.network(NETWORK_INTERFACE_CELLULAR);
-#endif // HAL_PLATFORM_CELLULAR
-
-#if HAL_PLATFORM_WIFI
-        config.network(NETWORK_INTERFACE_WIFI_STA);
-#endif // HAL_PLATFORM_WIFI
-
-#if HAL_PLATFORM_MESH
-        config.network(NETWORK_INTERFACE_MESH);
-#endif // HAL_PLATFORM_MESH
-
-#if HAL_PLATFORM_ETHERNET
-        config.network(NETWORK_INTERFACE_ETHERNET);
-#endif // HAL_PLATFORM_ETHERNET
-    }
-
-    if (param & SYSTEM_SLEEP_FLAG_NO_WAIT) {
-        config.wait(SystemSleepWait::NO_WAIT);
-    }
-
-    if (!(param & SYSTEM_SLEEP_FLAG_DISABLE_WKP_PIN)) {
-        config.gpio(WKP, RISING);
-    }
-
-    hal_wakeup_source_base_t* wakeupSource = nullptr;
-    int ret = system_sleep_ext(config.halConfig(), &wakeupSource, nullptr);
-    CHECK(ret);
-
-    ret = 0; // 0 for RTC wakeup reason.
-    if (wakeupSource) {
-        if (wakeupSource->type == HAL_WAKEUP_SOURCE_TYPE_GPIO) {
-            for (size_t i = 0; i < pins_count; i++) {
-                if (pins[i] == reinterpret_cast<hal_wakeup_source_gpio_t*>(wakeupSource)->pin) {
-                    ret = i + 1;
-                    break;
-                }
-            }
-        }
-        free(wakeupSource);
-    }
-    return ret;
-#else
-    network_connect_cancel(0, 1, 0, 0);
-    return system_sleep_pin_impl(pins, pins_count, modes, modes_count, seconds, param, reserved);
-#endif // HAL_PLATFORM_SLEEP20
 }
